prevent compiler from optimizing some calls. Fix scan_34 baseline, some other problems related to sums use int64_t rather than doubles to prevent overflows

+++ drivers/cpp/benchmarks/reduce/29_reduce_sum_of_min_of_pairs/baseline.hpp
@@ -10,8 +10,8 @@
    input: x=[3, 4, 0, 2, 3], y=[2, 5, 3, 1, 7]
    output: 10
 */
-double NO_INLINE correctSumOfMinimumElements(std::vector<double> const& x, std::vector<double> const& y) {
-    std::vector<double> z;
+int64_t NO_INLINE correctSumOfMinimumElements(std::vector<int64_t> const& x, std::vector<int64_t> const& y) {
+    std::vector<int64_t> z;
     z.resize(x.size());
     std::transform(x.begin(), x.end(), y.begin(), z.begin(), [] (const auto &a, const auto &b) {
         return std::min(a, b);
