+++ drivers/cpp/benchmarks/scan/30_scan_prefix_sum/baseline.hpp
@@ -9,6 +9,6 @@
    input: [1, 7, 4, 6, 6, 2]
    output: [1, 8, 12, 18, 24, 26]
 */
-void NO_INLINE correctPrefixSum(std::vector<double> const& x, std::vector<double> &output) {
+void NO_INLINE correctPrefixSum(std::vector<int64_t> const& x, std::vector<int64_t> &output) {
     std::inclusive_scan(x.begin(), x.end(), output.begin());
 }
